+++ HollowEngine/HollowEngine/src/Hollow/Graphics/ElementArrayBuffer.h
@@ -11,7 +11,7 @@ namespace Hollow {
 		HOLLOW_API ElementArrayBuffer();
 		HOLLOW_API ~ElementArrayBuffer();
 		 
-		HOLLOW_API void AddData(const void* data, unsigned int count, unsigned int size);
+		HOLLOW_API void AddData(const void* data, size_t count, size_t size);
 		HOLLOW_API void Bind() const;
 		HOLLOW_API void Unbind() const;
 		HOLLOW_API inline unsigned int GetCount() const { return mCount; }
